delete function

+++ app/views.py
@@ -126,14 +126,28 @@ def updateRoom(request , pk):
     context = {'form' : form }
     return render(request , 'app/room_form.html' , context )
 
-
+@login_required(login_url='login')
 def deleteRoom(request , pk ):
     room = Room.objects.get(id = pk)
+
+    if request.user != room.host :
+        return HttpResponse('You are not allow here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request , 'app/delete.html' , {'obj' : room } ) 
 
 
+@login_required(login_url='login')
+def deleteMessage(request , pk ):
+    message = Message.objects.get(id = pk)
 
+    if request.user != message.user :
+        return HttpResponse('You are not allow here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request , 'app/delete.html' , {'obj' : message } )
